+++ support/fixtures/factories/index.js
@@ -20,6 +20,7 @@ require('./contxtUserApplication');
 require('./contxtUserFavoriteApplication');
 require('./contxtUserPermissions');
 require('./contxtUserRole');
+require('./contxtUserStack');
 require('./costCenter');
 require('./costCenterFacility');
 require('./defaultAudiences');